Reorg local logical optimizer rules
